refactor(llm): reorganize llm package into clients/models/frameworks
Moves three files into directories that reflect their conceptual role:

- nemoguardrails/llm/clients/openai_chat_model.py
  -> nemoguardrails/llm/models/openai_chat.py

  It implements the LLMModel protocol and converts between our canonical
  types and wire-format dicts. That's a consumer of the client, not a
  client. Keeping it under clients/ conflated transport (BaseClient,
  OpenAICompatibleClient) with protocol surface (Model). Future
  AnthropicChatModel / OpenAIResponsesModel now have an obvious home.

- nemoguardrails/llm/default_framework.py
  -> nemoguardrails/llm/frameworks/default.py

  default was a sibling of frameworks.py, which read as "registry and
  default are co-equal." They're not — default is one implementation of
  the protocol the registry dispatches to. Placing it inside the
  frameworks/ package signals that relationship.

- nemoguardrails/llm/frameworks.py
  -> nemoguardrails/llm/frameworks/registry.py

  The module is mostly registry functions (get_framework,
  set_default_framework, _reset_frameworks). registry beats the default
  frameworks.py/__init__.py naming because there's no base class here
  that would justify a base.py.

The LLMFramework protocol stays in nemoguardrails/types.py (widely
imported; out of scope for this move). LangChainFramework stays in
integrations/langchain/ (moving it back into llm/ would undo the
decoupling work).

A frameworks/__init__.py re-exports the public API (register_framework,
get_framework, set_default_framework, get_default_framework,
_reset_frameworks) so callers using `from nemoguardrails.llm.frameworks
import ...` keep working unchanged.

Tests updated to import from the new paths. No behavior changes.
